Move ext/common -> app/common
