+++ flashlight/app/asr/tutorial/FinetuneCTC.cpp
@@ -27,7 +27,7 @@
 #include "flashlight/ext/common/DistributedUtils.h"
 #include "flashlight/ext/common/SequentialBuilder.h"
 #include "flashlight/ext/common/Serializer.h"
-#include "flashlight/ext/common/Runtime.h"
+#include "flashlight/app/common/Runtime.h"
 #include "flashlight/ext/plugin/ModulePlugin.h"
 #include "flashlight/fl/contrib/contrib.h"
 #include "flashlight/fl/flashlight.h"
@@ -37,7 +37,7 @@
 
 using fl::ext::Serializer;
 using fl::ext::afToVector;
-using fl::ext::getRunFile;
+using fl::app::getRunFile;
 using fl::lib::fileExists;
 using fl::lib::format;
 using fl::lib::getCurrentDate;
