+++ flashlight/app/lm/CMakeLists.txt
@@ -18,6 +18,7 @@ target_link_libraries(
   ${gflags_LIBRARIES}
   flashlight
   fl-libraries
+  flashlight-app-common
   )
 
 target_include_directories(
