+++ flashlight/ext/common/CMakeLists.txt
@@ -5,5 +5,4 @@ target_sources(
   PRIVATE
   ${CMAKE_CURRENT_LIST_DIR}/SequentialBuilder.cpp
   ${CMAKE_CURRENT_LIST_DIR}/DistributedUtils.cpp
-  ${CMAKE_CURRENT_LIST_DIR}/Runtime.cpp
   )
